// convex/_generated/api.ts
// This file is generated by Convex and will be replaced
// For development purposes, we'll create a minimal implementation

import { AnyApi } from "convex/server"

export const api: AnyApi = {
  files: {
    generateUploadUrl: {
      _type: "mutation",
    },
    saveFile: {
      _type: "mutation",
    },
    getFile: {
      _type: "query",
    },
    getFileUrl: {
      _type: "query",
    },
    getUserFiles: {
      _type: "query",
    },
    getChatFiles: {
      _type: "query",
    },
    deleteFile: {
      _type: "mutation",
    },
    searchFiles: {
      _type: "query",
    },
    updateFile: {
      _type: "mutation",
    },
  },
} as any